Refactor and small fixes in Quote type
This breaks quote.rs into smaller submodules quote.rs, quoteheader.rs,
and sigdata.rs.
